requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
brotli==1.2.0
certifi==2025.11.12
click==8.3.0
diskcache==5.6.3
exceptiongroup==1.3.0
fastapi==0.121.1
ffmpy==1.0.0
filelock==3.20.0
fsspec==2025.10.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
Jinja2==3.1.6
llama_cpp_python==0.3.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.2.6
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
ruff==0.14.4
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
websockets==15.0.1
